+++ spec/Command/Runnable/ServerController/FlushCommandSpec.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace spec\Zlikavac32\BeanstalkdLibBundle\Command\Runnable\ServerController;
 
 use Ds\Map;
+use Ds\Set;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Prophecy\Argument\Token\TokenInterface;
@@ -14,6 +15,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
 use Zlikavac32\BeanstalkdLib\Client;
+use Zlikavac32\BeanstalkdLib\JobState;
 use Zlikavac32\BeanstalkdLib\TubeHandle;
 use Zlikavac32\BeanstalkdLibBundle\Command\Runnable\ServerController\FlushCommand;
 
@@ -46,10 +48,13 @@ class FlushCommandSpec extends ObjectBehavior
             ->toArray()
             ->shouldReturn([
                 'flush',
+                'flush -b',
+                'flush -d',
                 'flush -f',
                 'flush -f <TUBE-NAME>',
                 'flush -f bar',
                 'flush -f foo',
+                'flush -r',
                 'flush <TUBE-NAME>',
                 'flush bar',
                 'flush foo',
@@ -79,9 +84,9 @@ class FlushCommandSpec extends ObjectBehavior
     ): void {
         $client->tube('foo')->willReturn($tubeHandle);
 
-        $tubeHandle->flush()->shouldBeCalled();
+        $tubeHandle->flush(new Set([JobState::READY(), JobState::BURIED(), JobState::DELAYED()]))->shouldBeCalled();
 
-        $this->run(['-f' => 1, 'tube-name' => 'foo'], $input, $helperSet, $output);
+        $this->run(['-f' => 1, '-r' => null, '-d' => null, '-b' => null, 'tube-name' => 'foo'], $input, $helperSet, $output);
     }
 
     public function it_should_flush_all_tubes_on_force(
@@ -101,10 +106,10 @@ class FlushCommandSpec extends ObjectBehavior
         $client->tube('foo')->willReturn($fooTubeHandle);
         $client->tube('bar')->willReturn($barTubeHandle);
 
-        $fooTubeHandle->flush()->shouldBeCalled();
-        $barTubeHandle->flush()->shouldBeCalled();
+        $fooTubeHandle->flush(new Set([JobState::READY(), JobState::BURIED(), JobState::DELAYED()]))->shouldBeCalled();
+        $barTubeHandle->flush(new Set([JobState::READY(), JobState::BURIED(), JobState::DELAYED()]))->shouldBeCalled();
 
-        $this->run(['-f' => 1, 'tube-name' => null], $input, $helperSet, $output);
+        $this->run(['-f' => 1, '-r' => null, '-d' => null, '-b' => null, 'tube-name' => null], $input, $helperSet, $output);
     }
 
     public function it_should_flush_tube_without_force_when_confirmed(
@@ -123,9 +128,9 @@ class FlushCommandSpec extends ObjectBehavior
 
         $client->tube('foo')->willReturn($tubeHandle);
 
-        $tubeHandle->flush()->shouldBeCalled();
+        $tubeHandle->flush(new Set([JobState::READY(), JobState::BURIED(), JobState::DELAYED()]))->shouldBeCalled();
 
-        $this->run(['-f' => null, 'tube-name' => 'foo'], $input, $helperSet, $output);
+        $this->run(['-f' => null, '-r' => null, '-d' => null, '-b' => null, 'tube-name' => 'foo'], $input, $helperSet, $output);
     }
 
     public function it_should_flush_all_tubes_without_force_when_confirmed(
@@ -152,10 +157,10 @@ class FlushCommandSpec extends ObjectBehavior
         $questionHelper->ask($input, $output, $this->confirmQuestionToken(2))
             ->willReturn(true);
 
-        $fooTubeHandle->flush()->shouldBeCalled();
-        $barTubeHandle->flush()->shouldBeCalled();
+        $fooTubeHandle->flush(new Set([JobState::READY(), JobState::BURIED(), JobState::DELAYED()]))->shouldBeCalled();
+        $barTubeHandle->flush(new Set([JobState::READY(), JobState::BURIED(), JobState::DELAYED()]))->shouldBeCalled();
 
-        $this->run(['-f' => null, 'tube-name' => null], $input, $helperSet, $output);
+        $this->run(['-f' => null, '-r' => null, '-d' => null, '-b' => null, 'tube-name' => null], $input, $helperSet, $output);
     }
 
     public function it_should_not_flush_tube_without_force_when_not_confirmed(
@@ -173,7 +178,7 @@ class FlushCommandSpec extends ObjectBehavior
 
         $client->tube(Argument::any())->shouldNotBeCalled();
 
-        $this->run(['-f' => null, 'tube-name' => 'foo'], $input, $helperSet, $output);
+        $this->run(['-f' => null, '-r' => null, '-d' => null, '-b' => null, 'tube-name' => 'foo'], $input, $helperSet, $output);
     }
 
     public function it_should_not_flush_all_tubes_without_force_when_not_confirmed(
@@ -200,7 +205,26 @@ class FlushCommandSpec extends ObjectBehavior
 
         $client->tube(Argument::any())->shouldNotBeCalled();
 
-        $this->run(['-f' => null, 'tube-name' => null], $input, $helperSet, $output);
+        $this->run(['-f' => null, '-r' => null, '-d' => null, '-b' => null, 'tube-name' => null], $input, $helperSet, $output);
+    }
+
+    public function it_should_flush_only_listed_states(
+        Client $client,
+        TubeHandle $tubeHandle,
+        InputInterface $input,
+        HelperSet $helperSet,
+        OutputInterface $output
+    ): void {
+        $client->tubes()
+               ->willReturn(new Map([
+                   'foo' => $tubeHandle->getWrappedObject(),
+               ]));
+
+        $client->tube('foo')->willReturn($tubeHandle);
+
+        $tubeHandle->flush(new Set([JobState::DELAYED(), JobState::BURIED()]))->shouldBeCalled();
+
+        $this->run(['-f' => 1, '-r' => null, '-d' => 1, '-b' => 1, 'tube-name' => null], $input, $helperSet, $output);
     }
 
     private function confirmQuestionToken(int $count): TokenInterface